+++ examples/spot/auto_invest/queryOnetimeTransactionStatus.js
@@ -1,0 +1,12 @@
+const Spot = require('../../../src/spot')
+
+const apiKey = ''
+const apiSecret = ''
+const client = new Spot(apiKey, apiSecret)
+
+client.queryOnetimeTransactionStatus(12345, {
+  requestId: 'TR12354859',
+  recvWindow: 5000
+})
+  .then(response => client.logger.log(response.data))
+  .catch(error => client.logger.error(error))