+++ src/APIBase.js
@@ -2,10 +2,12 @@
 
 const crypto = require('crypto')
 const { removeEmptyValue, buildQueryString, createRequest, defaultLogger } = require('./helpers/utils')
+const ConnectorClientError = require('./error/connectorClientError')
+const PrivateKeyAlgo = require('./helpers/privateKeyAlgo')
 
 class APIBase {
   constructor (options) {
-    const { apiKey, apiSecret, baseURL, logger, timeout, proxy, httpsAgent, privateKey, privateKeyPassphrase, wsURL } = options
+    const { apiKey, apiSecret, baseURL, logger, timeout, proxy, httpsAgent, privateKey, privateKeyPassphrase, privateKeyAlgo, wsURL } = options
 
     this.apiKey = apiKey
     this.apiSecret = apiSecret
@@ -17,6 +19,7 @@ class APIBase {
     this.logger = logger || defaultLogger
     this.privateKey = privateKey || ''
     this.privateKeyPassphrase = privateKeyPassphrase || ''
+    this.privateKeyAlgo = privateKeyAlgo || PrivateKeyAlgo.RSA
     this.wsURL = wsURL
   }
 
@@ -49,13 +52,20 @@ class APIBase {
         .update(queryString)
         .digest('hex')
     } else {
-      signature = crypto
-        .createSign('RSA-SHA256')
-        .update(queryString)
-        .sign({
+      if (this.privateKeyAlgo === PrivateKeyAlgo.RSA) {
+        signature = crypto.sign('RSA-SHA256', Buffer.from(queryString), {
           key: this.privateKey,
           passphrase: this.privateKeyPassphrase
-        }, 'base64')
+        }).toString('base64')
+      } else if (this.privateKeyAlgo === PrivateKeyAlgo.ED25519) {
+        signature = crypto.sign(null, Buffer.from(queryString), {
+          key: this.privateKey,
+          padding: crypto.constants.RSA_PKCS1_PSS_PADDING,
+          saltLength: crypto.constants.RSA_PSS_SALTLEN_DIGEST
+        }).toString('base64')
+      } else {
+        throw new ConnectorClientError("privateKeyAlgo must be either 'RSA' or 'ED25519'")
+      }
       signature = encodeURIComponent(signature)
     }