+++ CHANGELOG.md
@@ -1,5 +1,124 @@
 # Changelog
 
+## 3.4.1 - 2024-08-19
+### Updated
+- Updated dependencies
+- Upgraded deprecated spot endpoint `POST /api/v3/order/oco` to `POST /api/v3/orderList/oco`
+- Upgraded deprecated Websocket API endpoint `orderList.place` to `orderList.place.oco`
+
+## 3.4.0 - 2024-06-19
+### Changed
+- Updated dependencies
+- Resolved multiple websocket stream request issue
+
+## 3.3.0 - 2024-04-04
+
+### Added
+- Add Simple Earn endpoints:
+  - `GET /sapi/v1/simple-earn/flexible/list` to query available Simple Earn flexible product list
+  - `GET /sapi/v1/simple-earn/locked/list` to query available Simple Earn locked product list
+  - `POST /sapi/v1/simple-earn/flexible/subscribe` to subscribe to a flexible product
+  - `POST /sapi/v1/simple-earn/locked/subscribe` to subscribe to a locked product
+  - `POST /sapi/v1/simple-earn/flexible/redeem` to redeem a flexible product
+  - `POST /sapi/v1/simple-earn/locked/redeem` to redeem a locked product
+  - `GET /sapi/v1/simple-earn/flexible/position` to get a flexible product position
+  - `GET /sapi/v1/simple-earn/locked/position` to get a locked product position
+  - `GET /sapi/v1/simple-earn/account ` to get a simple account balances
+  - `GET /sapi/v1/simple-earn/flexible/history/subscriptionRecord` to get flexible subscription records
+  - `GET /sapi/v1/simple-earn/locked/history/subscriptionRecord ` to get locked subscription records
+  - `GET /sapi/v1/simple-earn/flexible/history/redemptionRecord ` to retrieve flexible redemption records
+  - `GET /sapi/v1/simple-earn/locked/history/redemptionRecord ` to retrieve locked redemption records
+  - `GET /sapi/v1/simple-earn/flexible/history/rewardsRecord ` to get flexible rewards history
+  - `GET /sapi/v1/simple-earn/locked/history/rewardsRecord ` to get locked rewards history
+  - `POST /sapi/v1/simple-earn/flexible/setAutoSubscribe` to set an auto-subscription to a flexible product
+  - `POST /sapi/v1/simple-earn/locked/setAutoSubscribe` to set an auto-subscription to a locked product
+  - `GET /sapi/v1/simple-earn/flexible/personalLeftQuota` to get flexible personal left quota
+  - `GET /sapi/v1/simple-earn/locked/personalLeftQuota` to get locked personal left quota
+  - `GET /sapi/v1/simple-earn/flexible/subscriptionPreview` to get flexible subscription preview
+  - `GET /sapi/v1/simple-earn/locked/subscriptionPreview` to get locked subscription previews
+  - `GET /sapi/v1/simple-earn/flexible/history/rateHistory` to get a rate history
+  - `GET /sapi/v1/simple-earn/flexible/history/collateralRecord` to get collateral records
+
+### Changed
+- Update dependencies
+
+## 3.2.0 - 2024-01-23
+
+### Changed
+- Updated `follow-redirects` dependency
+
+### Removed
+- Deprecated Savings endpoints:
+  - `GET /sapi/v1/lending/daily/product/list`
+  - `GET /sapi/v1/lending/daily/userLeftQuota`
+  - `POST /sapi/v1/lending/daily/purchase`
+  - `GET /sapi/v1/lending/daily/userRedemptionQuota`
+  - `POST /sapi/v1/lending/daily/redeem`
+  - `GET /sapi/v1/lending/daily/token/position`
+  - `GET /sapi/v1/lending/project/list`
+  - `POST /sapi/v1/lending/customizedFixed/purchase`
+  - `GET /sapi/v1/lending/union/account`
+  - `GET /sapi/v1/lending/union/purchaseRecord`
+  - `GET /sapi/v1/lending/union/redemptionRecord`
+  - `GET /sapi/v1/lending/union/interestHistory`
+  - `GET /sapi/v1/lending/project/position/list`
+
+- Deprecated Staking endpoints:
+  - `GET /sapi/v1/staking/productList`
+  - `POST /sapi/v1/staking/purchase`
+  - `POST /sapi/v1/staking/redeem`
+  - `GET /sapi/v1/staking/position`
+  - `GET /sapi/v1/staking/stakingRecord`
+  - `POST /sapi/v1/staking/setAutoStaking`
+  - `GET /sapi/v1/staking/personalLeftQuota`
+
+- Deprecated BSwap endpoints:
+  - `GET /sapi/v1/bswap/pools`
+  - `GET /sapi/v1/bswap/liquidity`
+  - `POST /sapi/v1/bswap/liquidityAdd`
+  - `POST /sapi/v1/bswap/liquidityRemove`
+  - `GET /sapi/v1/bswap/liquidityOps`
+  - `GET /sapi/v1/bswap/quote`
+  - `POST /sapi/v1/bswap/swap`
+  - `GET /sapi/v1/bswap/swap`
+  - `GET /sapi/v1/bswap/poolConfigure`
+  - `GET /sapi/v1/bswap/addLiquidityPreview`
+  - `GET /sapi/v1/bswap/removeLiquidityPreview`
+  - `GET /sapi/v1/bswap/unclaimedRewards`
+  - `POST /sapi/v1/bswap/claimRewards`
+  - `GET /sapi/v1/bswap/claimedHistory`
+
+## 3.1.0 - 2023-12-18
+
+### Added
+- Add Auto Invest endpoints:
+  - `GET /sapi/v1/margin/available-inventory` Query margin available inventory
+  - `POST /sapi/v1/margin/manual-liquidation` Margin manual liquidation
+
+### Changed
+- Update dependencies
+
+### Removed
+- `GET /sapi/v1/futures/loan/borrow/history`
+- `GET /sapi/v1/futures/loan/repay/history`
+- `GET /sapi/v2/futures/loan/wallet`
+- `GET /sapi/v1/futures/loan/adjustCollateral/history`
+- `GET /sapi/v1/futures/loan/liquidationHistory`
+- `GET /sapi/v1/futures/loan/interestHistory`
+
+## 3.0.0 - 2023-10-20
+
+### Changed
+- Update README #142
+- Minor changes #143
+- Update package
+- Update Nodejs versions in pipeline
+
+## 3.0.0-rc.2 - 2023-05-03
+
+### Added
+- Support Ed25519 key signature
+
 ## v3.0.0-rc1 - 2023-04-20
 
 ### Added